nursingace/requirements.txt
anyascii==0.3.1
asgiref==3.5.0
backports.zoneinfo==0.2.1; python_version < '3.9'
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
django-autoslug==1.9.8
django-ckeditor==6.4.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==2.0.0
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.5.1
django==4.0.4
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3; python_version >= '3'
l18n==2021.3
openpyxl==3.0.9
pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1; python_version >= '3.0'
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
urllib3==1.26.9
wagtail-metadata==3.5.0
wagtail==2.16.2
webencodings==0.5.1
willow==1.4.1
xlrd==2.0.1
xlsxwriter==3.0.3
xlwt==1.3.0
